+++ crypto-ingestor/Cargo.toml
@@ -26,7 +26,7 @@ config = "0.13"
 
 rust_decimal = "1"
 thiserror = "1"
-metrics = "0.21"
+metrics_core = { package = "metrics", version = "0.21" }
 
 [profile.release]
 opt-level = 3
