+++ crypto-ingestor/src/agents/coinbase/mod.rs
@@ -8,7 +8,7 @@ use crate::{
     config::Settings,
     error::IngestorError,
     http_client,
-    metrics::{ERRORS, RECONNECTS, TRADES_RECEIVED},
+    telemetry::{ERRORS, RECONNECTS, TRADES_RECEIVED},
 };
 use canonicalizer::CanonicalService;
