+++ crypto-ingestor/src/lib.rs
@@ -3,4 +3,4 @@ pub mod agents;
 pub mod config;
 pub mod http_client;
 pub mod error;
-pub mod metrics;
+pub mod telemetry;
