+++ crypto-ingestor/src/telemetry.rs
